+++ ninchatsdk/src/main/java/com/ninchat/sdk/adapters/NinchatMessageAdapter.java
@@ -30,6 +30,7 @@ import com.ninchat.sdk.NinchatSessionManager;
 import com.ninchat.sdk.R;
 import com.ninchat.sdk.activities.NinchatChatActivity;
 import com.ninchat.sdk.activities.NinchatMediaActivity;
+import com.ninchat.sdk.helper.EndlessRecyclerViewScrollListener;
 import com.ninchat.sdk.models.NinchatFile;
 import com.ninchat.sdk.models.NinchatMessage;
 import com.ninchat.sdk.models.NinchatUser;
@@ -320,31 +321,7 @@ public final class NinchatMessageAdapter extends RecyclerView.Adapter<NinchatMes
         }
     }
 
-    protected class ScrollListener extends RecyclerView.OnScrollListener {
-
-        private boolean updateData = false;
-        private int index;
-        private boolean updated;
-        private boolean removed;
-
-        protected void setData(final int index, final boolean updated, final boolean removed) {
-            updateData = true;
-            this.index = index;
-            this.updated = updated;
-            this.removed = removed;
-        }
-
-        @Override
-        public void onScrollStateChanged(@NonNull RecyclerView recyclerView, int newState) {
-            if (recyclerView.getScrollState() == RecyclerView.SCROLL_STATE_IDLE && updateData) {
-                messagesUpdated(index, updated, removed);
-                updateData = false;
-            }
-            super.onScrollStateChanged(recyclerView, newState);
-        }
-    }
-
-    protected ScrollListener scrollListener = new ScrollListener();
+    protected EndlessRecyclerViewScrollListener scrollListener = new EndlessRecyclerViewScrollListener((index, updated, removed) -> messagesUpdated(index, updated, removed));
 
     protected static final SimpleDateFormat TIMESTAMP_FORMATTER = new SimpleDateFormat("HH:mm", new Locale("fi-FI"));
